requirements.txt
torch 
torchvision
torchaudio
ultralytics
pyyaml
seaborn
scikit-learn